backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pyvips==2.2.1
pdf2image==1.16.3
Pillow==10.1.0
PyJWT==2.8.0
bcrypt==4.1.2
